+++ R/tune_race_anova.R
@@ -20,6 +20,10 @@
 #'  tuning parameter candidates. An integer denotes the number of candidate
 #'  parameter sets to be created automatically.
 #' @param metrics A [yardstick::metric_set()] or `NULL`.
+#' @param eval_time A numeric vector of time points where dynamic event time
+#' metrics should be computed (e.g. the time-dependent ROC curve, etc). The
+#' values must be non-negative and should probably be no greater than the
+#' largest event time in the training set (See Details below).
 #' @param control An object used to modify the tuning process. See
 #'  [control_race()] for more details.
 #' @param ... Not currently used.
@@ -31,7 +35,7 @@
 #'
 #' Racing methods are efficient approaches to grid search. Initially, the
 #'  function evaluates all tuning parameters on a small initial set of
-#'  resamples. The `burn_in` argument of `control_race()` sets the number of
+#'  resamples. The `burn_in` argument of [control_race()] sets the number of
 #'  initial resamples.
 #'
 #' The performance statistics from these resamples are analyzed to determine
@@ -46,12 +50,23 @@
 #' This function determines statistical significance using a repeated measures ANOVA
 #'  model where the performance statistic (e.g., RMSE, accuracy, etc.) is the
 #'  outcome data and the random effect is due to resamples. The
-#'  `control_race()` function contains are parameter for the significance cutoff
+#'  [control_race()] function contains are parameter for the significance cutoff
 #'  applied to the ANOVA results as well as other relevant arguments.
 #'
 #' There is benefit to using racing methods in conjunction with parallel
 #'  processing. The following section shows a benchmark of results for one
 #'  dataset and model.
+#'
+#' ## Censored regression models
+#'
+#' With dynamic performance metrics (e.g. Brier or ROC curves), performance is
+#' calculated for every value of `eval_time` but the _first_ evaluation time
+#' given by the user (e.g., `eval_time[1]`) is analyzed during racing.
+#'
+#' Also, values of `eval_time` should be less than the largest observed event
+#' time in the training data. For many non-parametric models, the results beyond
+#' the largest time corresponding to an event are constant (or `NA`).
+#'
 #' @return An object with primary class `tune_race` in the same standard format
 #' as objects produced by [tune::tune_grid()].
 #' @includeRmd  man/rmd/anova-benchmark.md details
@@ -103,99 +118,142 @@ tune_race_anova.default <- function(object, ...) {
     "The first argument to [tune_race_anova()] should be either ",
     "a model or workflow."
   )
-  rlang::abort(msg)
+  cli::cli_abort(msg)
 }
 
 #' @export
-tune_race_anova.recipe <- function(object, model, resamples, ..., param_info = NULL,
-                                   grid = 10, metrics = NULL,
-                                   control = control_race()) {
-  tune::empty_ellipses(...)
+tune_race_anova.recipe <-
+  function(object,
+           model,
+           resamples,
+           ...,
+           param_info = NULL,
+           grid = 10,
+           metrics = NULL,
+           eval_time = NULL,
+           control = control_race()) {
+    tune::empty_ellipses(...)
 
-  control <- parsnip::condense_control(control, control_race())
+    control <- parsnip::condense_control(control, control_race())
 
-  tune_race_anova(
-    model,
-    preprocessor = object, resamples = resamples,
-    param_info = param_info, grid = grid,
-    metrics = metrics, control = control
-  )
-}
+    tune_race_anova(
+      model,
+      preprocessor = object,
+      resamples = resamples,
+      param_info = param_info,
+      grid = grid,
+      metrics = metrics,
+      eval_time = eval_time,
+      control = control
+    )
+  }
 
 #' @export
-tune_race_anova.formula <- function(formula, model, resamples, ..., param_info = NULL,
-                                    grid = 10, metrics = NULL,
-                                    control = control_race()) {
-  tune::empty_ellipses(...)
+tune_race_anova.formula <-
+  function(formula,
+           model,
+           resamples,
+           ...,
+           param_info = NULL,
+           grid = 10,
+           metrics = NULL,
+           eval_time = NULL,
+           control = control_race()) {
+    tune::empty_ellipses(...)
 
-  control <- parsnip::condense_control(control, control_race())
+    control <- parsnip::condense_control(control, control_race())
 
-  tune_race_anova(
-    model,
-    preprocessor = formula, resamples = resamples,
-    param_info = param_info, grid = grid,
-    metrics = metrics, control = control
-  )
-}
+    tune_race_anova(
+      model,
+      preprocessor = formula,
+      resamples = resamples,
+      param_info = param_info,
+      grid = grid,
+      metrics = metrics,
+      eval_time = eval_time,
+      control = control
+    )
+  }
 
 #' @export
 #' @rdname tune_race_anova
-tune_race_anova.model_spec <- function(object, preprocessor, resamples, ...,
-                                       param_info = NULL, grid = 10, metrics = NULL,
-                                       control = control_race()) {
-  if (rlang::is_missing(preprocessor) || !tune::is_preprocessor(preprocessor)) {
-    rlang::abort(paste(
-      "To tune a model spec, you must preprocess",
-      "with a formula, recipe, or variable specification"
-    ))
+tune_race_anova.model_spec <-
+  function(object,
+           preprocessor,
+           resamples,
+           ...,
+           param_info = NULL,
+           grid = 10,
+           metrics = NULL,
+           eval_time = NULL,
+           control = control_race()) {
+    if (rlang::is_missing(preprocessor) || !tune::is_preprocessor(preprocessor)) {
+      cli::cli_abort(
+        "To tune a model spec, you must preprocess with a formula, recipe, \\
+        or variable specification."
+      )
+    }
+
+    tune::empty_ellipses(...)
+
+    control <- parsnip::condense_control(control, control_race())
+
+    wflow <- workflows::add_model(workflows::workflow(), object)
+
+    if (tune::is_recipe(preprocessor)) {
+      wflow <- workflows::add_recipe(wflow, preprocessor)
+    } else if (rlang::is_formula(preprocessor)) {
+      wflow <- workflows::add_formula(wflow, preprocessor)
+    }
+
+    tune_race_anova_workflow(
+      wflow,
+      resamples = resamples,
+      grid = grid,
+      metrics = metrics,
+      eval_time = eval_time,
+      param_info = param_info,
+      control = control
+    )
   }
-
-  tune::empty_ellipses(...)
-
-  control <- parsnip::condense_control(control, control_race())
-
-  wflow <- workflows::add_model(workflows::workflow(), object)
-
-  if (tune::is_recipe(preprocessor)) {
-    wflow <- workflows::add_recipe(wflow, preprocessor)
-  } else if (rlang::is_formula(preprocessor)) {
-    wflow <- workflows::add_formula(wflow, preprocessor)
-  }
-
-  tune_race_anova_workflow(
-    wflow,
-    resamples = resamples,
-    grid = grid,
-    metrics = metrics,
-    param_info = param_info,
-    control = control
-  )
-}
 
 #' @export
 #' @rdname tune_race_anova
-tune_race_anova.workflow <- function(object, resamples, ..., param_info = NULL,
-                                     grid = 10, metrics = NULL,
-                                     control = control_race()) {
-  tune::empty_ellipses(...)
+tune_race_anova.workflow <-
+  function(object,
+           resamples,
+           ...,
+           param_info = NULL,
+           grid = 10,
+           metrics = NULL,
+           eval_time = NULL,
+           control = control_race()) {
+    tune::empty_ellipses(...)
 
-  control <- parsnip::condense_control(control, control_race())
+    control <- parsnip::condense_control(control, control_race())
 
-  tune_race_anova_workflow(
-    object,
-    resamples = resamples,
-    grid = grid,
-    metrics = metrics,
-    param_info = param_info,
-    control = control
-  )
-}
+    tune_race_anova_workflow(
+      object,
+      resamples = resamples,
+      grid = grid,
+      metrics = metrics,
+      eval_time = eval_time,
+      param_info = param_info,
+      control = control
+    )
+  }
 
 ## -----------------------------------------------------------------------------
 
 tune_race_anova_workflow <-
-  function(object, resamples, param_info = NULL, grid = 10, metrics = NULL,
-           control = control_race()) {
+  function(object,
+           resamples,
+           param_info = NULL,
+           grid = 10,
+           metrics = NULL,
+           eval_time = NULL,
+           control = control_race(),
+           call = caller_env()) {
     rlang::check_installed("lme4")
 
     tune::initialize_catalog(control = control)
@@ -210,7 +268,20 @@ tune_race_anova_workflow <-
     check_num_resamples(B, min_rs)
     tmp_resamples <- restore_rset(resamples, 1:min_rs)
 
-    control$pkgs <- c(tune::required_pkgs(object), "workflows", "tidyr", "rlang")
+    metrics <- tune::check_metrics_arg(metrics, object, call = call)
+    eval_time <- tune::check_eval_time_arg(eval_time, metrics, call = call)
+
+    control$pkgs <- c(control$pkgs, tune::required_pkgs(object), "workflows", "tidyr", "rlang")
+
+    if (control$verbose_elim) {
+      tune_cols <- tune::get_tune_colors()
+      msg <- tune_cols$message$info(
+        paste0(cli::symbol$info,
+               " Evaluating against the initial {min_rs} burn-in resamples.")
+      )
+
+      cli::cli_inform(msg)
+    }
 
     grid_control <- parsnip::condense_control(control, tune::control_grid())
     res <-
@@ -220,29 +291,26 @@ tune_race_anova_workflow <-
         param_info = param_info,
         grid = grid,
         metrics = metrics,
+        eval_time = eval_time,
         control = grid_control
       )
 
     param_names <- tune::.get_tune_parameter_names(res)
-    metrics <- tune::.get_tune_metrics(res)
-    analysis_metric <- names(attr(metrics, "metrics"))[1]
-    analysis_max <- attr(attr(metrics, "metrics")[[1]], "direction") == "maximize"
 
-    cols <- tune::get_tune_colors()
-    if (control$verbose_elim) {
-      msg <-
-        paste(
-          "Racing will", ifelse(analysis_max, "maximize", "minimize"),
-          "the", analysis_metric, "metric."
-        )
-      rlang::inform(cols$message$info(paste0(cli::symbol$info, " ", msg)))
-      if (control$randomize) {
-        msg <- "Resamples are analyzed in a random order."
-        rlang::inform(cols$message$info(paste0(cli::symbol$info, " ", msg)))
-      }
-    }
+    opt_metric <- tune::first_metric(metrics)
+    opt_metric_name <- opt_metric$metric
+    maximize <- opt_metric$direction == "maximize"
 
-    filters_results <- test_parameters_gls(res, control$alpha)
+    opt_metric_time <- tune::first_eval_time(
+      metrics,
+      metric = opt_metric_name,
+      eval_time = eval_time,
+      call = call
+    )
+
+    racing_obj_log(opt_metric_name, opt_metric$direction, control, opt_metric_time)
+
+    filters_results <- test_parameters_gls(res, control$alpha, opt_metric_time)
     n_grid <- nrow(filters_results)
 
     log_final <- TRUE
@@ -258,11 +326,11 @@ tune_race_anova_workflow <-
 
       if (nrow(new_grid) > 1) {
         tmp_resamples <- restore_rset(resamples, rs)
-        log_racing(control, filters_results, res$splits, n_grid, analysis_metric)
+        log_racing(control, filters_results, res$splits, n_grid, opt_metric_name)
       } else {
         tmp_resamples <- restore_rset(resamples, rs:B)
         if (log_final) {
-          log_racing(control, filters_results, res$splits, n_grid, analysis_metric)
+          log_racing(control, filters_results, res$splits, n_grid, opt_metric_name)
         }
         log_final <- FALSE
       }
@@ -275,15 +343,16 @@ tune_race_anova_workflow <-
           param_info = param_info,
           grid = new_grid,
           metrics = metrics,
+          eval_time = eval_time,
           control = grid_control
         )
 
-      res <- restore_tune(res, tmp_res)
+      res <- restore_tune(res, tmp_res, opt_metric_time)
 
       if (nrow(new_grid) > 1) {
-        filters_results <- test_parameters_gls(res, control$alpha)
+        filters_results <- test_parameters_gls(res, control$alpha, opt_metric_time)
         if (sum(filters_results$pass) == 2 & num_ties >= control$num_ties) {
-          filters_results <- tie_breaker(res, control)
+          filters_results <- tie_breaker(res, control, eval_time = opt_metric_time)
         }
       } else {
         # Depending on the value of control$parallel_over we don't need to do
@@ -302,7 +371,7 @@ tune_race_anova_workflow <-
 
 check_num_resamples <- function(B, min_rs) {
   if (B <= min_rs) {
-    rlang::abort(
+    cli::cli_abort(
       paste0("The number of resamples (", B, ") needs to be more than the ",
              "number of burn-in resamples (", min_rs, ") set by the control ",
              "function `control_race()`."),