+++ tests/testthat/test-anova-overall.R
@@ -1,6 +1,9 @@
 
 test_that("formula interface", {
   skip_on_cran()
+  skip_if_not_installed("Matrix", "1.6-2")
+  skip_if_not_installed("lme4", "1.1-35.1")
+
   expect_snapshot({
     set.seed(1)
     res <- f_wflow %>%
@@ -10,14 +13,18 @@ test_that("formula interface", {
       )
   })
   expect_equal(class(res), c("tune_race", "tune_results", "tbl_df", "tbl", "data.frame"))
-  expect_true(nrow(collect_metrics(res)) < nrow(grid_mod) * 2)
+  expect_true(nrow(collect_metrics(res)) < nrow(grid_mod) * 3)
   expect_equal(res, .Last.tune.result)
+  expect_null(.get_tune_eval_times(res))
+  expect_null(.get_tune_eval_time_target(res))
 })
 
 # ------------------------------------------------------------------------------
 
 test_that("recipe interface", {
   skip_on_cran()
+  skip_if_not_installed("Matrix", "1.6-2")
+  skip_if_not_installed("lme4", "1.1-35.1")
   expect_silent({
     set.seed(1)
     res <- rec_wflow %>%
@@ -27,7 +34,7 @@ test_that("recipe interface", {
       )
   })
   expect_equal(class(res), c("tune_race", "tune_results", "tbl_df", "tbl", "data.frame"))
-  expect_true(nrow(collect_metrics(res)) < nrow(grid_mod) * 2)
+  expect_true(nrow(collect_metrics(res)) < nrow(grid_mod) * 3)
   expect_equal(res, .Last.tune.result)
 })
 
@@ -35,6 +42,9 @@ test_that("recipe interface", {
 
 test_that("variable interface", {
   skip_on_cran()
+  skip_if_not_installed("Matrix", "1.6-2")
+  skip_if_not_installed("lme4", "1.1-35.1")
+
   expect_silent({
     set.seed(1)
     res <- var_wflow %>%
@@ -44,7 +54,7 @@ test_that("variable interface", {
       )
   })
   expect_equal(class(res), c("tune_race", "tune_results", "tbl_df", "tbl", "data.frame"))
-  expect_true(nrow(collect_metrics(res))  < nrow(grid_mod) * 2)
+  expect_true(nrow(collect_metrics(res))  < nrow(grid_mod) * 3)
   expect_equal(res, .Last.tune.result)
 })
 
@@ -52,6 +62,8 @@ test_that("variable interface", {
 
 
 test_that("too few resamples", {
+  skip_if_not_installed("Matrix", "1.6-2")
+  skip_if_not_installed("lme4", "1.1-35.1")
 
   rs <- rsample::vfold_cv(modeldata::cells, v = 2)
   expect_snapshot_error(